Stations docs: fixing OWL def, plus refactoring in CpmetaFetcher.scala
